Split the phase instance tests into their own live class
The two instance tests reach the shared DARLING_TEST_PG store, which made
StoreCopyPhaseTests a live class and tripped
LivePostgresCollectionHygieneTests. That pin's preferred remedy is this
shape: a mostly-pure class keeps its purity and the live tests move to
their own ...LivePostgresTests. Putting the attribute on
StoreCopyPhaseTests instead would serialize seven source-text pins
against every other live class for nothing.

The vacuity guards travel with the tests: the start case asserts the row
loop was not reached, the data case asserts Begin did return. The start
case needs no store and stays with the data case anyway -- they are a
test and its negative control, and they share the probe plumbing.

Both phase counts now spell the assignment \s*=\s* rather than with
literal single spaces. copyPhase=StoreCopyPhase.Start; is valid C# and
evaded every literal form.
